+++ sdk/communication/azure-communication-email/azure/communication/email/_shared/models.py
@@ -28,7 +28,7 @@ class CommunicationIdentifierKind(str, Enum, metaclass=DeprecatedEnumMeta):
     COMMUNICATION_USER = "communication_user"
     PHONE_NUMBER = "phone_number"
     MICROSOFT_TEAMS_USER = "microsoft_teams_user"
-    MICROSOFT_BOT = "microsoft_bot"
+    MICROSOFT_TEAMS_APP = "microsoft_teams_app"
 
 
 class CommunicationCloudEnvironment(str, Enum, metaclass=CaseInsensitiveEnumMeta):
@@ -56,12 +56,9 @@ class CommunicationIdentifier(Protocol):
 CommunicationUserProperties = TypedDict("CommunicationUserProperties", {"id": str})
 
 PHONE_NUMBER_PREFIX = "4:"
-BOT_PREFIX = "28:"
-BOT_PUBLIC_CLOUD_PREFIX = "28:orgid:"
-BOT_DOD_CLOUD_PREFIX = "28:dod:"
-BOT_DOD_CLOUD_GLOBAL_PREFIX = "28:dod-global:"
-BOT_GCCH_CLOUD_PREFIX = "28:gcch:"
-BOT_GCCH_CLOUD_GLOBAL_PREFIX = "28:gcch-global:"
+TEAMS_APP_PUBLIC_CLOUD_PREFIX = "28:orgid:"
+TEAMS_APP_DOD_CLOUD_PREFIX = "28:dod:"
+TEAMS_APP_GCCH_CLOUD_PREFIX = "28:gcch:"
 TEAMS_USER_ANONYMOUS_PREFIX = "8:teamsvisitor:"
 TEAMS_USER_PUBLIC_CLOUD_PREFIX = "8:orgid:"
 TEAMS_USER_DOD_CLOUD_PREFIX = "8:dod:"
@@ -242,74 +239,60 @@ def _microsoft_teams_user_raw_id(identifier: MicrosoftTeamsUserIdentifier) -> st
     return f"{TEAMS_USER_PUBLIC_CLOUD_PREFIX}{user_id}"
 
 
-_MicrosoftBotProperties = TypedDict(
-    "MicrosoftBotProperties",
+MicrosoftTeamsAppProperties = TypedDict(
+    "MicrosoftTeamsAppProperties",
     {
-        "bot_id": str,
-        "is_resource_account_configured": bool,
+        "app_id": str,
         "cloud": Union[CommunicationCloudEnvironment, str],
     },
 )
 
 
-class _MicrosoftBotIdentifier:
-    """Represents an identifier for a Microsoft bot.
+class MicrosoftTeamsAppIdentifier:
+    """Represents an identifier for a Microsoft Teams app.
 
     :ivar str raw_id: Optional raw ID of the identifier.
     :ivar kind: The type of identifier.
     :vartype kind: str or CommunicationIdentifierKind
-    :ivar MicrosoftBotProperties: The properties of the identifier.
+    :ivar MicrosoftTeamsAppProperties: The properties of the identifier.
      The keys in this mapping include:
-        - `bot_id`(str): The id of the Microsoft bot.
-        - `is_resource_account_configured` (bool): Set this to false if the bot is global.
-        The default is `true` for tennantized bots.
+        - `app_id`(str): The id of the Microsoft Teams app.
         - `cloud` (str): Cloud environment that this identifier belongs to.
 
-    :param str bot_id: Microsoft bot id.
-    :keyword bool is_resource_account_configured: `False` if the identifier is global.
-    Default value is `True` for tennantzed bots.
-    :keyword cloud: Cloud environment that the bot belongs to. Default value is `PUBLIC`.
+    :param str app_id: Microsoft Teams app id.
+    :keyword cloud: Cloud environment that the teams app belongs to. Default value is `PUBLIC`.
     :paramtype cloud: str or ~azure.communication.chat.CommunicationCloudEnvironment
     """
 
-    kind = CommunicationIdentifierKind.MICROSOFT_BOT
+    kind = CommunicationIdentifierKind.MICROSOFT_TEAMS_APP
 
-    def __init__(self, bot_id: str, **kwargs: Any) -> None:
+    def __init__(self, app_id: str, **kwargs: Any) -> None:
         self.raw_id = kwargs.get("raw_id")
-        self.properties = _MicrosoftBotProperties(
-            bot_id=bot_id,
-            is_resource_account_configured=kwargs.get(
-                "is_resource_account_configured", True
-            ),
+        self.properties = MicrosoftTeamsAppProperties(
+            app_id=app_id,
             cloud=kwargs.get("cloud") or CommunicationCloudEnvironment.PUBLIC,
         )
         if self.raw_id is None:
-            self.raw_id = _microsoft_bot_raw_id(self)
+            self.raw_id = _microsoft_teams_app_raw_id(self)
 
     def __eq__(self, other):
         try:
-            return self.raw_id == _microsoft_bot_raw_id(other)
+            return self.raw_id == _microsoft_teams_app_raw_id(other)
         except Exception:  # pylint: disable=broad-except
             return False
 
 
-def _microsoft_bot_raw_id(identifier: _MicrosoftBotIdentifier) -> str:  # pylint: disable=too-many-return-statements
+def _microsoft_teams_app_raw_id(identifier: MicrosoftTeamsAppIdentifier) -> str:  # pylint: disable=too-many-return-statements
     if identifier.raw_id:
         return str(identifier.raw_id)
-    bot_id = identifier.properties["bot_id"]
+    app_id = identifier.properties["app_id"]
     cloud = identifier.properties["cloud"]
-    if identifier.properties["is_resource_account_configured"] is False:
-        if cloud == CommunicationCloudEnvironment.DOD:
-            return f"{BOT_DOD_CLOUD_GLOBAL_PREFIX}{bot_id}"
-        if cloud == CommunicationCloudEnvironment.GCCH:
-            return f"{BOT_GCCH_CLOUD_GLOBAL_PREFIX}{bot_id}"
-        return f"{BOT_PREFIX}{bot_id}"
 
     if cloud == CommunicationCloudEnvironment.DOD:
-        return f"{BOT_DOD_CLOUD_PREFIX}{bot_id}"
+        return f"{TEAMS_APP_DOD_CLOUD_PREFIX}{app_id}"
     if cloud == CommunicationCloudEnvironment.GCCH:
-        return f"{BOT_GCCH_CLOUD_PREFIX}{bot_id}"
-    return f"{BOT_PUBLIC_CLOUD_PREFIX}{bot_id}"
+        return f"{TEAMS_APP_GCCH_CLOUD_PREFIX}{app_id}"
+    return f"{TEAMS_APP_PUBLIC_CLOUD_PREFIX}{app_id}"
 
 
 def identifier_from_raw_id(raw_id: str) -> CommunicationIdentifier:  # pylint: disable=too-many-return-statements
@@ -365,4 +348,22 @@ def identifier_from_raw_id(raw_id: str) -> CommunicationIdentifier:  # pylint: d
         SPOOL_USER_PREFIX,
     ]:
         return CommunicationUserIdentifier(id=raw_id, raw_id=raw_id)  # type: ignore
+    if prefix == TEAMS_APP_PUBLIC_CLOUD_PREFIX:
+        return MicrosoftTeamsAppIdentifier(
+            app_id=suffix,
+            cloud=CommunicationCloudEnvironment.PUBLIC,
+            raw_id=raw_id,
+        )  # type: ignore
+    if prefix == TEAMS_APP_GCCH_CLOUD_PREFIX:
+        return MicrosoftTeamsAppIdentifier(
+            app_id=suffix,
+            cloud=CommunicationCloudEnvironment.GCCH,
+            raw_id=raw_id,
+        )  # type: ignore
+    if prefix == TEAMS_APP_DOD_CLOUD_PREFIX:
+        return MicrosoftTeamsAppIdentifier(
+            app_id=suffix,
+            cloud=CommunicationCloudEnvironment.DOD,
+            raw_id=raw_id,
+        )  # type: ignore
     return UnknownIdentifier(identifier=raw_id)  # type: ignore